[AGC] Recognize agc in triple parsing code
Added a new triple for the Apollo Guidance Computer (AGC) and allow the
triple parsing to recognise this triple and its corresponding
properties.

The target has 16-bit pointers but not all of this space is usable.

The minimum addressable unit is a 16-bit (including parity) word, and
double words are arranged in a big-endian manner.

The target does have an assembler/linker already that uses a custom
executable format, however we will be using ELF as we are focussing on
the compilation portion.