[AGC] Switch to using pseudo instructions for reg operands until emission
Since we have enforced a distinction between memory and registers
manually on top of instructions which can actually be used for all
memory locations, we want to try and keep instructions which enforce
registers as operands for use in code generation, such that we can use
register allocation and distinguish registers from memory. The variants
that can take arbitrary addresses can be kept for manual expansion of
pseudo instructions for more complex sequences, or - as implemented in
this patch - as the last stage in emission of the pseudo instructions
which take register operands.

This requires some slightly fiddly modification of the code emitter and
disassembler to map pseudo instructions to non-pseudo for encoding
purposes, and vice-versa for getting register operands from decoding
instructions which don't report register operands.