[AGC] Add basic CodeGen implementation
This patch adds the necessary files to implement a CodeGen target for
AGC. Only a very limited set of ALU operations are implemented so far.

The calling convention implemented here is not complete. The next
requirement is to implement call lowering and an emulated stack,
including frame handling.

Register copies required an extensive amount of implementation to cover
the variety of accumulator/non-accumulator cases, and ensuring we don't
overwrite the accumulator erroneously.

Tests for basic ALU operations and argument lowering are added, along
with support for UpdateTestChecks for AGC.