[AGC] Work required to support loading constants from a pool
This patch includes an implementation for the use of loads from
linker-defined constant pool addresses to load constants into the
accumulator register. Thus it requires a large amount of boilerplate and
changes to get the mappings between fixups, relocations, custom
expressions etc. working.

The constant pool address is represented by a relocation R_AGC_CPI12,
with an offset argument of the actual constant itself - thus giving the
linker the information needed to process constants defined in an object
file. In assembly the expression modifier used for this is %cpi(N).

Zeros are simply loaded from the fixed zero register R7.