[AGC] Implement basic support for stack and register spilling
This patch is an accumulation of modifications required to support
situations where the register allocator would need to evict registers,
and use either the stack or copies to other registers. To do this, we
implement support for storing and loading to/from a frame index, support
for epilogue/prologue insertion and for expanding frame indices into a
register plus offset.