[AGC] Add basic support for calls to functions
This patch adds the necessary lowering to produce calls, placement of
input arguments and extraction of output arguments. For now the callee
must be a global address, and in order to resolve this address we must
insert more relocations such that the linker can insert the appropriate
bank selection bits and the appropriate lower bits of the address when
it is resolved. We need to then load these into reserved registers to
use in a function dispatch call. This is necessary to facilitate the
switching of bank configurations - which change the effective program
counter - but without ending up in an unknown location. Instead the
dispatch function is duplicated in each bank so that after the banks are
switched we will continue executing the same code.

Similar to the argument lowering we don't yet support stack arguments.
Tail calls are also not supported, and neither are varargs.