[AGC] Implement branches for AGC
This patch adds lowering and selection of conditional and unconditional
branches, using the TCF, BZF (branch-if-zero) and BZMF
(branch-if-zero-or-minus) instructions.

Normal comparisons between two variables are translated into comparisons
against zero such that the branch instructions can be used. Given that
there is no way to branch if not equal to zero with a single instruction
we need to use a pseudo instruction and expand it into a pattern using
the inverted branch.

Proper branch analysis still needs to be implemented in AGCInstrInfo to
improve code generation.