[AGC] Implement branch analysis to improve code generation of branches
This patch adds implementations for the analyzeBranch, removeBranch and
insertBranch hooks to allow generic code generation transformations to
work with AGC branches. We also switch the expansion of the PseudoBNZF
instruction to use these hooks as it is simply an inversion of the
branch condition and the target basic blocks.